+++ geotrek/tourism/tests/test_views.py
@@ -645,3 +645,15 @@ class TouristicEventViewSetTest(TestCase):
         response = self.client.get('/api/en/touristicevents.geojson', data={'ends_after': '2020-01-01'})
         geojson = json.loads(response.content)
         self.assertEqual(len(geojson['features']), 5)
+
+
+class TouristicContentCategoryViewSetTest(TestCase):
+    def test_get_categories(self):
+        """
+        Test category json serialization via api
+        """
+        nb_elements = 10
+        TouristicContentCategoryFactory.create_batch(nb_elements)
+        response = self.client.get(reverse('tourism:touristic_categories_json', kwargs={'lang': 'en'}))
+        json_response = json.loads(response.content)
+        self.assertEqual(len(json_response), nb_elements)