+++ geotrek/trekking/management/commands/sync_rando.py
@@ -17,7 +17,6 @@ from django.test.client import RequestFactory
 from django.utils import translation
 
 from mapentity.settings import app_settings as mapentity_settings
-
 from geotrek.altimetry.views import ElevationProfile, ElevationArea, serve_elevation_chart
 from geotrek.common import models as common_models
 from geotrek.trekking import models as trekking_models
@@ -182,6 +181,13 @@ class Command(BaseCommand):
             zipfile.write(fullname, name)
         if self.verbosity == '2':
             self.stdout.write(u"\x1b[3D\x1b[32mgenerated\x1b[0m")
+
+    def sync_json(self, lang, viewset, name, zipfile=None, params={}, **kwargs):
+        view = viewset.as_view()
+        name = os.path.join('api', lang, '{name}.json'.format(name=name))
+        if self.source:
+            params['source'] = ','.join(self.source)
+        self.sync_view(lang, view, name, params=params, zipfile=zipfile, **kwargs)
 
     def sync_geojson(self, lang, viewset, name, zipfile=None, params={}, **kwargs):
         view = viewset.as_view({'get': 'list'})