+++ docs/changelog.rst
@@ -10,6 +10,7 @@ CHANGELOG
 * Sync rando now exports only future events based on current date.
 * Sync rando now synchronizes touristic contents and events.
 * Sync rando now handles crashes when it calls django views.
+* Sync rando now synchronizes touristic content categories.
 
 **Bug fixes**